+++ jme-render/src/test/kotlin/info/laht/krender/jme/TestPrimitives.kt
@@ -1,7 +1,7 @@
 package info.laht.krender.jme
 
 import info.laht.krender.ColorConstants
-import org.joml.Vector3d
+import org.joml.Vector3f
 
 fun main() {
 
@@ -10,22 +10,22 @@ fun main() {
         init()
 
         createBox(1f, 1f, 1f).apply {
-            setTranslate(Vector3d(-2.0, 0.0, 0.0))
+            setTranslate(Vector3f(-2f, 0f, 0f))
             setColor(ColorConstants.bisque)
         }
 
         createSphere(0.5f).apply {
-            setTranslate(Vector3d(0.0, 0.0, 0.0))
+            setTranslate(Vector3f(0f, 0f, 0f))
             setColor(ColorConstants.firebrick)
         }
 
         createCapsule(0.5f, 1f).apply {
-            setTranslate(Vector3d(2.0, 0.0, 0.0))
+            setTranslate(Vector3f(2f, 0f, 0f))
             setColor(ColorConstants.antiquewhite)
         }
 
         createCylinder(0.5f, 1f).apply {
-            setTranslate(Vector3d(4.0, 0.0, 0.0))
+            setTranslate(Vector3f(4f, 0f, 0f))
             setColor(ColorConstants.rebeccapurple)
         }
